requirements.txt
flask
requests
beautifulsoup4
urllib3
gunicorn
python-telegram-bot
python-dotenv